% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big2utf8.R
\name{big2utf8}
\alias{big2utf8}
\title{convert file format from big5 to utf8}
\usage{
big2utf8(filename, newfile)
}
\description{
This function allows you to covert file from big5 to utf8.
}
\examples{
big2utf8(sourcefile,newfile)
}
\keyword{big5}
